Make it possible to asynchronously provide credentials
