+++ aws/rust-runtime/aws-auth/Cargo.toml
@@ -10,3 +10,7 @@ edition = "2018"
 [dependencies]
 smithy-http = { path = "../../../rust-runtime/smithy-http" }
 zeroize = "1.2.0"
+
+[dev-dependencies]
+http = "0.2.3"
+tokio = { version = "1.0", features = ["rt", "macros"] }
