+++ rust-runtime/smithy-http/src/middleware.rs
@@ -17,9 +17,29 @@ use bytes::{Buf, Bytes};
 use http::Response;
 use http_body::Body;
 use std::error::Error;
+use std::future::Future;
+use std::pin::Pin;
 use tracing::trace;
 
 type BoxError = Box<dyn Error + Send + Sync>;
+type BoxFuture<T> = Pin<Box<dyn Future<Output = T> + Send>>;
+
+/// [`AsyncMapRequest`] defines an asynchronous middleware that transforms an [`operation::Request`].
+///
+/// Typically, these middleware will read configuration from the `PropertyBag` and use it to
+/// augment the request.
+///
+/// Most fundamental middleware is expressed as `AsyncMapRequest`'s synchronous cousin, `MapRequest`,
+/// including signing & endpoint resolution. `AsyncMapRequest` is used for async credential
+/// retrieval (e.g., from AWS STS's AssumeRole operation).
+pub trait AsyncMapRequest {
+    type Error: Into<BoxError> + 'static;
+
+    fn apply(
+        &self,
+        request: operation::Request,
+    ) -> BoxFuture<Result<operation::Request, Self::Error>>;
+}
 
 /// [`MapRequest`] defines a synchronous middleware that transforms an [`operation::Request`].
 ///
